[BE] Fix: 결제 금액 반환값 수정(카드 결제)
long -> double

+++ bank/src/main/java/com/goofy/tunabank/v1/dto/card/CardPaymentResponseDto.java
@@ -18,6 +18,6 @@ public class CardPaymentResponseDto {
   private LocalDateTime paymentAt;
 
   private String currencyCode;
-  private Long paymentBalance;
+  private Double paymentBalance;
 
 }
